Transformation into Chaincode
The code from the Oracle DB save calls ‘rest_pkg.save_qr_scan’, which writes the data to the oracle database after it has been validated. After writing the user input data to the database, the same dataset is transformed into a format that the Chaincode is expecting, which is given with this code ("Transformation into Chaincode"). The code is then calling a RESTful API webservice on the Hyperledger Fabric Blockchain instance by executing: ‘apex_web_service.make_rest_request()’.